requirements.txt
aws-cdk-lib
constructs
boto3
jsonschema>=4